[Features]   - When local pdf report doesn't exist, it will be automatically downloaded if available in cloud storage, and if network is available.
[Bugfixes and Improvements]
  - Moved pdf generation into a background service.
