+++ app/src/main/java/com/skichrome/oc/easyvgp/model/local/LocalVgpListSource.kt
@@ -1,6 +1,5 @@
 package com.skichrome.oc.easyvgp.model.local
 
-import android.net.Uri
 import com.skichrome.oc.easyvgp.model.Results
 import com.skichrome.oc.easyvgp.model.Results.Error
 import com.skichrome.oc.easyvgp.model.Results.Success
@@ -10,6 +9,7 @@ import com.skichrome.oc.easyvgp.util.NotImplementedException
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
+import java.io.File
 
 class LocalVgpListSource(
     private val userDao: UserDao,
@@ -33,19 +33,7 @@ class LocalVgpListSource(
         }
     }
 
-    override suspend fun getReportFromDate(date: Long): Results<List<Report>> = withContext(dispatchers) {
-        return@withContext try
-        {
-            val result = machineControlPointDataDao.getPreviouslyInsertedReport(date)
-            Success(result)
-        }
-        catch (e: Exception)
-        {
-            Error(e)
-        }
-    }
-
-    override suspend fun getReportExtrasFromId(id: Long): Results<MachineControlPointDataExtra> = withContext(dispatchers) {
+    override suspend fun getMachineCtrlPtExtraFromId(id: Long): Results<MachineControlPointDataExtra> = withContext(dispatchers) {
         return@withContext try
         {
             Success(machineCtrlPtExtraDao.getExtraFromId(id))
@@ -56,10 +44,10 @@ class LocalVgpListSource(
         }
     }
 
-    override suspend fun getUserFromId(id: Long): Results<UserAndCompany> = withContext(dispatchers) {
+    override suspend fun updateMachineCtrlPtExtra(extra: MachineControlPointDataExtra): Results<Int> = withContext(dispatchers) {
         return@withContext try
         {
-            Success(userDao.getUserFromId(id))
+            Success(machineCtrlPtExtraDao.update(extra))
         }
         catch (e: Exception)
         {
@@ -67,84 +55,6 @@ class LocalVgpListSource(
         }
     }
 
-    override suspend fun getCustomerFromId(customerId: Long): Results<Customer> = withContext(dispatchers) {
-        return@withContext try
-        {
-            Success(customerDao.getCustomerById(customerId))
-        }
-        catch (e: Exception)
-        {
-            Error(e)
-        }
-    }
-
-    override suspend fun getMachineFromId(machineId: Long): Results<Machine> = withContext(dispatchers) {
-        return@withContext try
-        {
-            Success(machineDao.getMachineById(machineId))
-        }
-        catch (e: Exception)
-        {
-            Error(e)
-        }
-    }
-
-    override suspend fun getMachineTypeFromId(machineTypeId: Long): Results<MachineType> = withContext(dispatchers) {
-        return@withContext try
-        {
-            Success(machineTypeDao.getMachineTypeFromId(machineTypeId))
-        }
-        catch (e: Exception)
-        {
-            Error(e)
-        }
-    }
-
-    override suspend fun updateMachine(machine: Machine): Results<Int> = withContext(dispatchers) {
-        return@withContext try
-        {
-            Success(machineDao.update(machine))
-        }
-        catch (e: Exception)
-        {
-            Error(e)
-        }
-    }
-
-    override suspend fun updateUser(user: User): Results<Int> = withContext(dispatchers) {
-        return@withContext try
-        {
-            Success(userDao.update(user))
-        }
-        catch (e: Exception)
-        {
-            Error(e)
-        }
-    }
-
-    override suspend fun updateCompany(company: Company): Results<Int> = withContext(dispatchers) {
-        return@withContext try
-        {
-            Success(companyDao.update(company))
-        }
-        catch (e: Exception)
-        {
-            Error(e)
-        }
-    }
-
-    override suspend fun uploadImageToStorage(userUid: String, localUri: Uri, remoteUri: Uri?, filePrefix: String): Results<Uri> =
-        Error(NotImplementedException("Method not available on local VGPList source"))
-
-    override suspend fun generateReport(
-        reportDate: Long,
-        user: UserAndCompany,
-        customer: Customer,
-        machine: Machine,
-        machineType: MachineType,
-        reports: List<Report>,
-        reportExtra: MachineControlPointDataExtra
-
-    ): Results<Boolean> =
+    override suspend fun downloadReportFromStorage(remotePath: String?, destinationFile: File): Results<File> =
         Error(NotImplementedException("Method not available on local VGPList source"))
 }
