+++ app/src/main/java/com/skichrome/oc/easyvgp/model/local/database/MachineControlPointData.kt
@@ -48,7 +48,7 @@ data class Report(
 @Dao
 interface MachineControlPointDataDao : BaseDao<MachineControlPointData>
 {
-    @Query("SELECT MachinesControlPointsDataExtras.report_date, MachinesControlPointsDataExtras.report_path_on_device, MachinesControlPointsDataExtras.machines_control_points_data_extras_id, MachinesControlPointsDataCrossRef.ctrl_point_data_id, MachinesControlPointsDataExtras.is_report_valid FROM MachinesControlPointsDataCrossRef LEFT JOIN ControlPointsData JOIN MachinesControlPointsDataExtras WHERE MachinesControlPointsDataCrossRef.ctrl_point_data_id == ControlPointsData.ctrl_point_data_id AND MachinesControlPointsDataCrossRef.machine_id == :id AND MachinesControlPointsDataExtras.machines_control_points_data_extras_id == MachinesControlPointsDataCrossRef.machine_ctrl_point_data_extras_reference")
+    @Query("SELECT MachinesControlPointsDataExtras.report_date, MachinesControlPointsDataExtras.report_path_on_device, MachinesControlPointsDataExtras.report_path_on_remote_storage, MachinesControlPointsDataExtras.machines_control_points_data_extras_id, MachinesControlPointsDataCrossRef.ctrl_point_data_id, MachinesControlPointsDataExtras.is_report_valid FROM MachinesControlPointsDataCrossRef LEFT JOIN ControlPointsData JOIN MachinesControlPointsDataExtras WHERE MachinesControlPointsDataCrossRef.ctrl_point_data_id == ControlPointsData.ctrl_point_data_id AND MachinesControlPointsDataCrossRef.machine_id == :id AND MachinesControlPointsDataExtras.machines_control_points_data_extras_id == MachinesControlPointsDataCrossRef.machine_ctrl_point_data_extras_reference")
     suspend fun getCtrlPointDataFromMachineId(id: Long): List<VgpListItem>
 
     @Query("SELECT * FROM MachinesControlPointsDataCrossRef JOIN ControlPointsData JOIN control_points JOIN MachinesControlPointsDataExtras WHERE MachinesControlPointsDataExtras.report_date == :reportDate AND MachinesControlPointsDataCrossRef.ctrl_point_data_id == ControlPointsData.ctrl_point_data_id AND ControlPointsData.ctrl_point_data_ctrl_point_ref == control_points.control_point_id AND MachinesControlPointsDataExtras.machines_control_points_data_extras_id == MachinesControlPointsDataCrossRef.machine_ctrl_point_data_extras_reference")
